Move interfaces to new namespaces
