+++ Firefly.Core/Services/Encryption/IAsymmetricalEncryption.cs
@@ -1,0 +1,9 @@
+﻿namespace Firefly.Core.Services.Encryption
+{
+    /// <summary>
+    /// Represents an interface for performing asymmetrical encryption
+    /// </summary>
+    public interface IAsymmetricalEncryption : IEncryptionService
+    {
+    }
+}
