+++ Firefly.Core/Services/Encryption/ISymmetricalEncryption.cs
@@ -1,0 +1,9 @@
+﻿namespace Firefly.Core.Services.Encryption
+{
+    /// <summary>
+    /// Represents an interface for performing symmetrical encryption
+    /// </summary>
+    public interface ISymmetricalEncryption : IEncryptionService
+    {
+    }
+}
