--- Firefly.Core/Services/IAsymmetricalEncryption.cs
@@ -1,6 +1,0 @@
-﻿namespace Firefly.Core.Services.Security
-{
-    public interface IAsymmetricalEncryption : IEncryptionService
-    {
-    }
-}
