--- Firefly.Core/Services/ISymmetricalEncryption.cs
@@ -1,6 +1,0 @@
-﻿namespace Firefly.Core.Services.Security
-{
-    public interface ISymmetricalEncryption : IEncryptionService
-    {
-    }
-}
